Simplify the table implementation
On the first pass I went overboard with caching and other complexities.
In practice it doesn't hurt to re-evaluate the values each time a new
expression is evaluated, so we'll do that for simplicity's sake.

visitSymbols now gets the ability to visit only a subset of symbols, while
still preserving the topological sort order.